Do unit conversions with f64s
Fixes #25

+++ common/src/units/defs.rs
@@ -23,17 +23,17 @@ marker_structs![
 
 impl<P: MetricPrefix> LengthUnit for Meter<P> {}
 impl<P: MetricPrefix> Unit for Meter<P> {
-    const FACTOR: f32 = P::FACTOR;
+    const FACTOR: f64 = P::FACTOR;
 }
 
 impl<P: MetricPrefix> TimeUnit for Second<P> {}
 impl<P: MetricPrefix> Unit for Second<P> {
-    const FACTOR: f32 = P::FACTOR;
+    const FACTOR: f64 = P::FACTOR;
 }
 
 impl TimeUnit for Minute {}
 impl Unit for Minute {
-    const FACTOR: f32 = 60.0;
+    const FACTOR: f64 = 60.0;
 }
 
 pub struct Kilo;
@@ -43,21 +43,21 @@ pub struct Milli;
 pub struct Micro;
 
 impl MetricPrefix for Kilo {
-    const FACTOR: f32 = 1e3;
+    const FACTOR: f64 = 1e3;
 }
 
 impl MetricPrefix for Base {
-    const FACTOR: f32 = 1e0;
+    const FACTOR: f64 = 1e0;
 }
 
 impl MetricPrefix for Centi {
-    const FACTOR: f32 = 1e-2;
+    const FACTOR: f64 = 1e-2;
 }
 
 impl MetricPrefix for Milli {
-    const FACTOR: f32 = 1e-3;
+    const FACTOR: f64 = 1e-3;
 }
 
 impl MetricPrefix for Micro {
-    const FACTOR: f32 = 1e-6;
+    const FACTOR: f64 = 1e-6;
 }
